Move Admin Page's DashStatsBottom and DashStats to separate files inside _inc/client/at-a-glance/stats
